Move UpdateToast to toasts folder
