+++ components/toasts/UpdateToast/index.scss
